Refactors Footer and updates PageContainer for consistency
Wraps Footer content inside PageContainer for layout consistency and simplifies component structure.

Renames and extends PageContainer's max-width for better reuse. Updates relevant imports across the application to reflect the new location. Enhances layout uniformity on the homepage by wrapping key components in PageContainer.
